pyproject.toml
[tool.poetry]
name = "rhizome-contract-explorer"
version = "0.1.0"
description = ""
authors = ["Brian Li <brian@brianli.com>"]
readme = "README.md"
packages = [{include = "rhizome_contract_explorer"}]

[tool.poetry.dependencies]
python = "3.10.8"
iconsdk = "^2.2.0"
fastapi = "^0.88.0"
python-dotenv = "^0.21.0"
uvicorn = "^0.20.0"
jinja2 = "^3.1.2"
python-multipart = "^0.0.5"
yarl = "^1.8.2"
pyyaml = "^6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
